Rename the JTI namespace in test cases
