+++ tests/ORM/Driver/SQLite/Inheritance/JTI/CompositePKTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\ORM\Tests\Driver\SQLite\Inheritance\JTI;
+
+class CompositePKTest extends \Cycle\ORM\Tests\Inheritance\JTI\CompositePKTest
+{
+    public const DRIVER = 'sqlite';
+}
