+++ tests/ORM/Driver/SQLite/Inheritance/JTI/Mapper/WithRelationsClasslessMapperTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\ORM\Tests\Driver\SQLite\Inheritance\JTI;
+
+class WithRelationsClasslessMapperTest extends \Cycle\ORM\Tests\Inheritance\JTI\Mapper\WithRelationsClasslessMapperTest
+{
+    public const DRIVER = 'sqlite';
+}
