+++ tests/ORM/Driver/SQLite/Inheritance/JTI/Mapper/WithRelationsPromiseMapperTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\ORM\Tests\Driver\SQLite\Inheritance\JTI\Mapper;
+
+class WithRelationsPromiseMapperTest extends \Cycle\ORM\Tests\Inheritance\JTI\Mapper\WithRelationsPromiseMapperTest
+{
+    public const DRIVER = 'sqlite';
+}
