+++ tests/ORM/Driver/SQLite/Inheritance/JTI/Mapper/WithRelationsStdMapperTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\ORM\Tests\Driver\SQLite\Inheritance\JTI\Mapper\Mapper;
+
+class WithRelationsStdMapperTest extends \Cycle\ORM\Tests\Inheritance\JTI\Mapper\WithRelationsStdMapperTest
+{
+    public const DRIVER = 'sqlite';
+}
