+++ tests/ORM/Driver/SQLite/Inheritance/JTI/SimpleCasesTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\ORM\Tests\Driver\SQLite\Inheritance\JTI;
+
+class SimpleCasesTest extends \Cycle\ORM\Tests\Inheritance\JTI\SimpleCasesTest
+{
+    public const DRIVER = 'sqlite';
+}
