+++ tests/ORM/Driver/SQLite/Inheritance/JTI/SingleTableTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\ORM\Tests\Driver\SQLite\Inheritance\JTI;
+
+class SingleTableTest extends \Cycle\ORM\Tests\Inheritance\JTI\SingleTableTest
+{
+    public const DRIVER = 'sqlite';
+}
