+++ tests/ORM/Driver/SQLite/Inheritance/JTI/WithRelationsTest.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Cycle\ORM\Tests\Driver\SQLite\Inheritance\JTI;
+
+class WithRelationsTest extends \Cycle\ORM\Tests\Inheritance\JTI\WithRelationsTest
+{
+    public const DRIVER = 'sqlite';
+}
