--- tests/ORM/Driver/SQLite/Relation/JTI/CompositePKTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Driver\SQLite\Relation\JTI;
-
-class CompositePKTest extends \Cycle\ORM\Tests\Relation\JTI\CompositePKTest
-{
-    public const DRIVER = 'sqlite';
-}
