--- tests/ORM/Driver/SQLite/Relation/JTI/Mapper/WithRelationsClasslessMapperTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Driver\SQLite\Relation\JTI;
-
-class WithRelationsClasslessMapperTest extends \Cycle\ORM\Tests\Relation\JTI\Mapper\WithRelationsClasslessMapperTest
-{
-    public const DRIVER = 'sqlite';
-}
