--- tests/ORM/Driver/SQLite/Relation/JTI/Mapper/WithRelationsPromiseMapperTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Driver\SQLite\Relation\JTI\Mapper;
-
-class WithRelationsPromiseMapperTest extends \Cycle\ORM\Tests\Relation\JTI\Mapper\WithRelationsPromiseMapperTest
-{
-    public const DRIVER = 'sqlite';
-}
