--- tests/ORM/Driver/SQLite/Relation/JTI/Mapper/WithRelationsStdMapperTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Driver\SQLite\Relation\JTI\Mapper\Mapper;
-
-class WithRelationsStdMapperTest extends \Cycle\ORM\Tests\Relation\JTI\Mapper\WithRelationsStdMapperTest
-{
-    public const DRIVER = 'sqlite';
-}
