--- tests/ORM/Driver/SQLite/Relation/JTI/SchemaTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Driver\SQLite\Relation\JTI;
-
-class SchemaTest extends \Cycle\ORM\Tests\Relation\JTI\SchemaTest
-{
-    public const DRIVER = 'sqlite';
-}
