--- tests/ORM/Driver/SQLite/Relation/JTI/SimpleCasesTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Driver\SQLite\Relation\JTI;
-
-class SimpleCasesTest extends \Cycle\ORM\Tests\Relation\JTI\SimpleCasesTest
-{
-    public const DRIVER = 'sqlite';
-}
