--- tests/ORM/Driver/SQLite/Relation/JTI/SingleTableTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Driver\SQLite\Relation\JTI;
-
-class SingleTableTest extends \Cycle\ORM\Tests\Relation\JTI\SingleTableTest
-{
-    public const DRIVER = 'sqlite';
-}
