--- tests/ORM/Driver/SQLite/Relation/JTI/WithRelationsTest.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Driver\SQLite\Relation\JTI;
-
-class WithRelationsTest extends \Cycle\ORM\Tests\Relation\JTI\WithRelationsTest
-{
-    public const DRIVER = 'sqlite';
-}
