+++ tests/ORM/Inheritance/JTI/CompositePKTest.php
@@ -2,32 +2,35 @@
 
 declare(strict_types=1);
 
-namespace Cycle\ORM\Tests\Relation\JTI;
+namespace Cycle\ORM\Tests\Inheritance\JTI;
 
 use Cycle\ORM\Heap\Heap;
 use Cycle\ORM\Mapper\Mapper;
 use Cycle\ORM\SchemaInterface;
 use Cycle\ORM\Select;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Employee;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Engineer;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Manager;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Programator;
+use Cycle\ORM\Tests\Inheritance\Fixture\Employee;
+use Cycle\ORM\Tests\Inheritance\Fixture\Engineer;
+use Cycle\ORM\Tests\Inheritance\Fixture\Manager;
+use Cycle\ORM\Tests\Inheritance\Fixture\Programator;
 
-abstract class SingleTableTest extends SimpleCasesTest
+abstract class CompositePKTest extends SimpleCasesTest
 {
     protected const
-        EMPLOYEE_1 = ['employee_id' => 1, 'name' => 'John', 'age' => 38],
-        EMPLOYEE_2 = ['employee_id' => 2, 'name' => 'Anton', 'age' => 35],
-        EMPLOYEE_3 = ['employee_id' => 3, 'name' => 'Kentarius', 'age' => 27],
-        EMPLOYEE_4 = ['employee_id' => 4, 'name' => 'Valeriy', 'age' => 32],
+        EMPLOYEE_1 = ['id' => 1, 'employee_id' => 2, 'name' => 'John', 'age' => 38],
+        EMPLOYEE_2 = ['id' => 2, 'employee_id' => 3, 'name' => 'Anton', 'age' => 35],
+        EMPLOYEE_3 = ['id' => 3, 'employee_id' => 4, 'name' => 'Kentarius', 'age' => 27],
+        EMPLOYEE_4 = ['id' => 4, 'employee_id' => 5, 'name' => 'Valeriy', 'age' => 32],
 
-        ENGINEER_2 = ['discriminator' => 'engineer', 'role_id' => 2, 'level' => 8, 'rank' => ''],
-        ENGINEER_4 = ['discriminator' => 'engineer', 'role_id' => 4, 'level' => 10, 'rank' => ''],
-        MANAGER_1  = ['discriminator' => 'manager', 'role_id' => 1, 'level' => 0, 'rank' => 'top'],
-        MANAGER_3  = ['discriminator' => 'manager', 'role_id' => 3, 'level' => 0, 'rank' => 'bottom'],
+        ENGINEER_2 = ['_type' => 'engineer', 'id' => 2, 'role_id' => 3, 'level' => 8, 'rank' => null],
+        ENGINEER_4 = ['_type' => 'engineer', 'id' => 4, 'role_id' => 5, 'level' => 10, 'rank' => null],
+        MANAGER_1 =  ['_type' => 'manager', 'id' => 1, 'role_id' => 2, 'level' => null, 'rank' => 'top'],
+        MANAGER_3 =  ['_type' => 'manager', 'id' => 3, 'role_id' => 4, 'level' => null, 'rank' => 'bottom'],
 
-        PROGRAMATOR_2 = ['subrole_id' => 2, 'language' => 'php'],
-        PROGRAMATOR_4 = ['subrole_id' => 4, 'language' => 'go'],
+        ENGINEER_2_PK = ['id' => 2, 'role_id' => 3],
+        PROGRAMATOR_2_PK = ['second_id' => 3, 'subrole_id' => 2],
+
+        PROGRAMATOR_2 = ['second_id' => 3, 'subrole_id' => 2, 'language' => 'php'],
+        PROGRAMATOR_4 = ['second_id' => 5, 'subrole_id' => 4, 'language' => 'go'],
 
         EMPLOYEE_1_LOADED = self::EMPLOYEE_1,
         EMPLOYEE_2_LOADED = self::EMPLOYEE_2,
@@ -45,7 +48,6 @@ abstract class SingleTableTest extends SimpleCasesTest
 
         EMPLOYEE_ALL_LOADED = [self::EMPLOYEE_1_LOADED, self::EMPLOYEE_2_LOADED, self::EMPLOYEE_3_LOADED, self::EMPLOYEE_4_LOADED],
         ENGINEER_ALL_LOADED = [self::ENGINEER_2_LOADED, self::ENGINEER_4_LOADED],
-        ROLES_ALL_LOADED = [self::EMPLOYEE_1_LOADED, self::ENGINEER_2_LOADED, self::EMPLOYEE_3_LOADED, self::ENGINEER_4_LOADED],
         PROGRAMATOR_ALL_LOADED = [self::PROGRAMATOR_2_LOADED, self::PROGRAMATOR_4_LOADED],
         MANAGER_ALL_LOADED = [self::MANAGER_1_LOADED, self::MANAGER_3_LOADED],
 
@@ -56,27 +58,35 @@ abstract class SingleTableTest extends SimpleCasesTest
     public function setUp(): void
     {
         JtiBaseTest::setUp();
-        $this->logger->hide();
+        // $this->logger->hide();
 
-        $this->makeTable('employee_table', [
-            'employee_id_column' => 'integer',
+        $this->makeTable('employee', [
+            'id' => 'integer',
+            'employee_id' => 'integer',
             'name_column' => 'string',
             'age' => 'integer,nullable',
-        ], pk: ['employee_id_column']);
-        $this->makeTable('role_table', [
-            'role_id_column' => 'integer,nullable',
-            'subrole_id_column' => 'integer,nullable',
-            'discriminator' => 'string,nullable',
+        ], pk: ['id', 'employee_id']);
+        $this->makeTable('role', [
+            'id' => 'integer',
+            'role_id' => 'integer',
+            '_type' => 'string',
             'level' => 'integer,nullable',
             'rank' => 'string,nullable',
-            'language' => 'string,nullable',
         ], fk: [
-            'role_id_column' => ['table' => 'employee_table', 'column' => 'employee_id_column'],
-            'subrole_id_column' => ['table' => 'role_table', 'column' => 'role_id_column'],
-        ]);
+            'id' => ['table' => 'employee', 'column' => 'id'],
+            'role_id' => ['table' => 'employee', 'column' => 'employee_id'],
+        ], pk: ['id', 'role_id']);
+        $this->makeTable('programator', [
+            'id' => 'integer',
+            'subrole_id' => 'integer',
+            'language' => 'string',
+        ], fk: [
+            'id' => ['table' => 'engineer', 'column' => 'role_id'],
+            'subrole_id' => ['table' => 'engineer', 'column' => 'id'],
+        ], pk: ['id', 'subrole_id']);
 
-        $this->getDatabase()->table('employee_table')->insertMultiple(
-            ['employee_id_column', 'name_column', 'age'],
+        $this->getDatabase()->table('employee')->insertMultiple(
+            ['id', 'employee_id', 'name_column', 'age'],
             [
                 self::EMPLOYEE_1,
                 self::EMPLOYEE_2,
@@ -84,8 +94,8 @@ abstract class SingleTableTest extends SimpleCasesTest
                 self::EMPLOYEE_4,
             ]
         );
-        $this->getDatabase()->table('role_table')->insertMultiple(
-            ['discriminator', 'role_id_column', 'level', 'rank'],
+        $this->getDatabase()->table('role')->insertMultiple(
+            ['_type', 'id', 'role_id', 'level', 'rank'],
             [
                 self::MANAGER_1,
                 self::ENGINEER_2,
@@ -93,13 +103,14 @@ abstract class SingleTableTest extends SimpleCasesTest
                 self::ENGINEER_4,
             ]
         );
-        $this->getDatabase()->table('role_table')->insertMultiple(
-            ['subrole_id_column', 'language'],
+        $this->getDatabase()->table('programator')->insertMultiple(
+            ['id', 'subrole_id', 'language'],
             [
                 self::PROGRAMATOR_2,
                 self::PROGRAMATOR_4,
             ]
         );
+
         $this->logger->display();
     }
 
@@ -110,14 +121,15 @@ abstract class SingleTableTest extends SimpleCasesTest
                 SchemaInterface::ROLE        => 'employee',
                 SchemaInterface::MAPPER      => Mapper::class,
                 SchemaInterface::DATABASE    => 'default',
-                SchemaInterface::TABLE       => 'employee_table',
-                SchemaInterface::PRIMARY_KEY => 'employee_id',
+                SchemaInterface::TABLE       => 'employee',
+                SchemaInterface::PRIMARY_KEY => ['id', 'employee_id'],
                 SchemaInterface::COLUMNS     => [
-                    'employee_id' => 'employee_id_column',
+                    'id',
+                    'employee_id',
                     'name' => 'name_column',
                     'age',
                 ],
-                SchemaInterface::TYPECAST    => ['employee_id' => 'int', 'age' => 'int'],
+                SchemaInterface::TYPECAST    => ['id' => 'int', 'employee_id' => 'int', 'age' => 'int'],
                 SchemaInterface::SCHEMA      => [],
                 SchemaInterface::RELATIONS   => [],
             ],
@@ -130,28 +142,28 @@ abstract class SingleTableTest extends SimpleCasesTest
             'role' => [
                 SchemaInterface::MAPPER      => Mapper::class,
                 SchemaInterface::DATABASE    => 'default',
-                SchemaInterface::TABLE       => 'role_table',
-                SchemaInterface::PARENT      => 'employee',
+                SchemaInterface::TABLE       => 'role',
+                SchemaInterface::PARENT      => Employee::class,
                 SchemaInterface::CHILDREN    => [
                     'engineer' => Engineer::class,
                     'manager'  => Manager::class,
                 ],
-                SchemaInterface::PRIMARY_KEY => 'role_id',
-                SchemaInterface::DISCRIMINATOR => 'discriminator',
-                SchemaInterface::COLUMNS     => ['role_id' => 'role_id_column', 'level', 'rank', 'discriminator'],
-                SchemaInterface::TYPECAST    => ['role_id' => 'int', 'level' => 'int'],
+                SchemaInterface::PRIMARY_KEY => ['id', 'role_id'],
+                SchemaInterface::COLUMNS     => ['id', 'role_id', 'level', 'rank', '_type'],
+                SchemaInterface::TYPECAST    => ['id' => 'int', 'role_id' => 'int', 'level' => 'int'],
                 SchemaInterface::SCHEMA      => [],
                 SchemaInterface::RELATIONS   => [],
             ],
             Programator::class => [
-                SchemaInterface::ROLE        => 'subrole',
+                SchemaInterface::ROLE        => 'programator',
                 SchemaInterface::MAPPER      => Mapper::class,
                 SchemaInterface::DATABASE    => 'default',
-                SchemaInterface::TABLE       => 'role_table',
+                SchemaInterface::TABLE       => 'programator',
                 SchemaInterface::PARENT      => Engineer::class,
-                SchemaInterface::PRIMARY_KEY => 'subrole_id',
-                SchemaInterface::COLUMNS     => ['subrole_id' => 'subrole_id_column', 'language'],
-                SchemaInterface::TYPECAST    => ['subrole_id' => 'int'],
+                SchemaInterface::PARENT_KEY  => ['role_id', 'id'],
+                SchemaInterface::PRIMARY_KEY => ['second_id', 'subrole_id'],
+                SchemaInterface::COLUMNS     => ['second_id' => 'id', 'subrole_id', 'language'],
+                SchemaInterface::TYPECAST    => ['second_id' => 'int', 'subrole_id' => 'int'],
                 SchemaInterface::SCHEMA      => [],
                 SchemaInterface::RELATIONS   => [],
             ],
@@ -162,31 +174,16 @@ abstract class SingleTableTest extends SimpleCasesTest
     {
         $selector = (new Select($this->orm, Engineer::class))
             // todo: this condition should be added automatically by STI
-            ->where('discriminator', '=', 'engineer');
+            ->where('_type', '=', 'engineer');
 
         $this->assertEquals(static::ENGINEER_ALL_LOADED, $selector->fetchData());
-    }
-
-    public function testFetchAllChildren(): void
-    {
-        /** @var Manager[]|Engineer[] $entities */
-        $entities = (new Select($this->orm, 'role'))->fetchAll();
-
-        $this->assertSame(1, $entities[0]->role_id);
-        $this->assertInstanceOf(Manager::class, $entities[0]);
-        $this->assertSame(2, $entities[1]->role_id);
-        $this->assertInstanceOf(Engineer::class, $entities[1]);
-        $this->assertSame(3, $entities[2]->role_id);
-        $this->assertInstanceOf(Manager::class, $entities[2]);
-        $this->assertSame(4, $entities[3]->role_id);
-        $this->assertInstanceOf(Engineer::class, $entities[3]);
     }
 
     public function testSelectEngineerDataFirst(): void
     {
         $selector = (new Select($this->orm, Engineer::class))
             // todo: this condition should be added automatically by STI
-            ->where('discriminator', '=', 'engineer')
+            ->where('_type', '=', 'engineer')
             ->limit(1);
 
         $this->assertEquals(static::ENGINEER_2_LOADED, $selector->fetchData()[0]);
@@ -196,7 +193,7 @@ abstract class SingleTableTest extends SimpleCasesTest
     {
         $selector = (new Select($this->orm, Manager::class))
             // todo: this condition should be added automatically by STI
-            ->where('discriminator', '=', 'manager');
+            ->where('_type', '=', 'manager');
 
         $this->assertEquals(static::MANAGER_ALL_LOADED, $selector->fetchData());
     }
@@ -205,7 +202,7 @@ abstract class SingleTableTest extends SimpleCasesTest
     {
         $selector = (new Select($this->orm, Manager::class))
             // todo: this condition should be added automatically by STI
-            ->where('discriminator', '=', 'manager')
+            ->where('_type', '=', 'manager')
             ->limit(1);
 
         $this->assertEquals(static::MANAGER_1_LOADED, $selector->fetchData()[0]);
@@ -214,6 +211,8 @@ abstract class SingleTableTest extends SimpleCasesTest
     public function testCreateProgramator(): void
     {
         $programator = new Programator();
+        $programator->id = 10;
+        $programator->employee_id = 11;
         $programator->name = 'Merlin';
         $programator->level = 50;
         $programator->language = 'VanillaJS';
@@ -226,17 +225,15 @@ abstract class SingleTableTest extends SimpleCasesTest
         $this->save($programator);
         $this->assertNumWrites(0);
 
-        $this->assertSame(5, $programator->employee_id);
-        $this->assertSame(5, $programator->role_id);
-        $this->assertSame(5, $programator->subrole_id);
-
         /** @var Programator $programator */
         $programator = (new Select($this->orm->withHeap(new Heap()), Programator::class))
-            ->wherePK($programator->subrole_id)
+            ->wherePK(['second_id' => 11, 'subrole_id' => 10])
             ->fetchOne();
-        $this->assertSame(5, $programator->employee_id);
-        $this->assertSame(5, $programator->role_id);
-        $this->assertSame(5, $programator->subrole_id);
+        $this->assertSame(10, $programator->id);
+        $this->assertSame(11, $programator->employee_id);
+        $this->assertSame(11, $programator->role_id);
+        $this->assertSame(10, $programator->subrole_id);
+        $this->assertSame(11, $programator->second_id);
         $this->assertSame('Merlin', $programator->name);
         $this->assertSame(50, $programator->level);
         $this->assertSame('VanillaJS', $programator->language);
