+++ tests/ORM/Inheritance/JTI/WithRelationsTest.php
@@ -2,18 +2,18 @@
 
 declare(strict_types=1);
 
-namespace Cycle\ORM\Tests\Relation\JTI;
+namespace Cycle\ORM\Tests\Inheritance\JTI;
 
 use Cycle\ORM\Mapper\Mapper;
 use Cycle\ORM\Relation;
 use Cycle\ORM\SchemaInterface;
 use Cycle\ORM\Select;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Book;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Employee;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Engineer;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Manager;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Programator;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Tool;
+use Cycle\ORM\Tests\Inheritance\Fixture\Book;
+use Cycle\ORM\Tests\Inheritance\Fixture\Employee;
+use Cycle\ORM\Tests\Inheritance\Fixture\Engineer;
+use Cycle\ORM\Tests\Inheritance\Fixture\Manager;
+use Cycle\ORM\Tests\Inheritance\Fixture\Programator;
+use Cycle\ORM\Tests\Inheritance\Fixture\Tool;
 use Cycle\ORM\Transaction;
 
 abstract class WithRelationsTest extends SimpleCasesTest
