--- tests/ORM/Relation/JTI/SchemaTest.php
@@ -1,190 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\ORM\Tests\Relation\JTI;
-
-use Cycle\ORM\Exception\SchemaException;
-use Cycle\ORM\Mapper\Mapper;
-use Cycle\ORM\Schema;
-use Cycle\ORM\SchemaInterface;
-use Cycle\ORM\Select;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Employee;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Engineer;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Manager;
-use Cycle\ORM\Tests\Relation\JTI\Fixture\Programator;
-
-abstract class SchemaTest extends JtiBaseTest
-{
-    public function setUp(): void
-    {
-        parent::setUp();
-
-        $this->makeTable('employee', [
-            'id'          => 'integer',
-            'name_column' => 'string',
-            'age'         => 'integer,nullable',
-        ], pk: ['id']);
-        $this->makeTable('engineer', [
-            'id'        => 'integer',
-            'level'     => 'integer',
-        ], fk: [
-            'id' => ['table' => 'employee', 'column' => 'id']
-        ], pk: ['id']);
-        $this->makeTable('programator', [
-            'id'        => 'integer',
-            'language' => 'string',
-        ], fk: [
-            'id' => ['table' => 'engineer', 'column' => 'id']
-        ], pk: ['id']);
-        $this->makeTable('manager', [
-            'id'        => 'integer',
-            'rank'      => 'string',
-        ], fk: [
-            'id' => ['table' => 'employee', 'column' => 'id']
-        ], pk: ['id']);
-
-        $this->getDatabase()->table('employee')->insertMultiple(
-            ['id', 'name_column', 'age'],
-            [
-                self::EMPLOYEE_1,
-                self::EMPLOYEE_2,
-                self::EMPLOYEE_3,
-                self::EMPLOYEE_4,
-            ]
-        );
-
-        $this->getDatabase()->table('engineer')->insertMultiple(
-            ['id', 'level'],
-            [
-                self::ENGINEER_2,
-                self::ENGINEER_4,
-            ]
-        );
-
-        $this->getDatabase()->table('programator')->insertMultiple(
-            ['id', 'language'],
-            [
-                self::PROGRAMATOR_2,
-                self::PROGRAMATOR_4,
-            ]
-        );
-
-        $this->getDatabase()->table('manager')->insertMultiple(
-            ['id', 'rank'],
-            [
-                self::MANAGER_1,
-                self::MANAGER_3,
-            ]
-        );
-    }
-
-    protected function getSchemaArray(): array
-    {
-        return [
-            Employee::class => [
-                Schema::ROLE        => 'employee',
-                Schema::MAPPER      => Mapper::class,
-                Schema::DATABASE    => 'default',
-                Schema::TABLE       => 'employee',
-                Schema::PRIMARY_KEY => 'id',
-                Schema::COLUMNS     => ['id', 'name' => 'name_column', 'age'],
-                Schema::TYPECAST    => ['id' => 'int', 'age' => 'int'],
-                Schema::SCHEMA      => [],
-                Schema::RELATIONS   => [],
-            ],
-            Engineer::class => [
-                Schema::ROLE        => 'engineer',
-                Schema::MAPPER      => Mapper::class,
-                Schema::DATABASE    => 'default',
-                Schema::TABLE       => 'engineer',
-                Schema::PARENT      => 'employee',
-                Schema::PRIMARY_KEY => 'id',
-                Schema::COLUMNS     => ['id', 'level'],
-                Schema::TYPECAST    => ['id' => 'int', 'level' => 'int'],
-                Schema::SCHEMA      => [],
-                Schema::RELATIONS   => [],
-            ],
-            Programator::class => [
-                Schema::ROLE        => 'programator',
-                Schema::MAPPER      => Mapper::class,
-                Schema::DATABASE    => 'default',
-                Schema::TABLE       => 'programator',
-                Schema::PARENT      => 'engineer',
-                Schema::PRIMARY_KEY => 'id',
-                Schema::COLUMNS     => ['id', 'language'],
-                Schema::TYPECAST    => ['id' => 'int'],
-                Schema::SCHEMA      => [],
-                Schema::RELATIONS   => [],
-            ],
-            Manager::class => [
-                Schema::ROLE        => 'manager',
-                Schema::MAPPER      => Mapper::class,
-                Schema::DATABASE    => 'default',
-                Schema::TABLE       => 'manager',
-                Schema::PARENT      => 'employee',
-                Schema::PRIMARY_KEY => 'id',
-                Schema::COLUMNS     => ['id', 'rank'],
-                Schema::TYPECAST    => ['id' => 'int'],
-                Schema::SCHEMA      => [],
-                Schema::RELATIONS   => [],
-            ],
-        ];
-    }
-
-    // todo: move this case to schema-checker
-    // /**
-    //  * @runInSeparateProcess
-    //  */
-    // public function testCyclicParents(): void
-    // {
-    //     set_time_limit(1);
-    //
-    //     $schema = [
-    //         Employee::class => [
-    //             SchemaInterface::ROLE        => 'employee',
-    //             SchemaInterface::MAPPER      => Mapper::class,
-    //             SchemaInterface::DATABASE    => 'default',
-    //             SchemaInterface::TABLE       => 'employee',
-    //             SchemaInterface::PARENT      => 'programator',
-    //             SchemaInterface::PRIMARY_KEY => 'id',
-    //             SchemaInterface::COLUMNS     => ['id', 'name' => 'name_column', 'age'],
-    //             SchemaInterface::TYPECAST    => ['id' => 'int', 'age' => 'int'],
-    //             SchemaInterface::SCHEMA      => [],
-    //             SchemaInterface::RELATIONS   => [],
-    //         ],
-    //         Engineer::class => [
-    //             SchemaInterface::ROLE        => 'engineer',
-    //             SchemaInterface::MAPPER      => Mapper::class,
-    //             SchemaInterface::DATABASE    => 'default',
-    //             SchemaInterface::TABLE       => 'engineer',
-    //             SchemaInterface::PARENT      => 'employee',
-    //             SchemaInterface::PRIMARY_KEY => 'id',
-    //             SchemaInterface::COLUMNS     => ['id', 'level'],
-    //             SchemaInterface::TYPECAST    => ['id' => 'int', 'level' => 'int'],
-    //             SchemaInterface::SCHEMA      => [],
-    //             SchemaInterface::RELATIONS   => [],
-    //         ],
-    //         Programator::class => [
-    //             SchemaInterface::ROLE        => 'programator',
-    //             SchemaInterface::MAPPER      => Mapper::class,
-    //             SchemaInterface::DATABASE    => 'default',
-    //             SchemaInterface::TABLE       => 'programator',
-    //             SchemaInterface::PARENT      => 'engineer',
-    //             SchemaInterface::PRIMARY_KEY => 'id',
-    //             SchemaInterface::COLUMNS     => ['id', 'language'],
-    //             SchemaInterface::TYPECAST    => ['id' => 'int'],
-    //             SchemaInterface::SCHEMA      => [],
-    //             SchemaInterface::RELATIONS   => [],
-    //         ],
-    //     ];
-    //     $this->orm = $this->orm->withSchema(new Schema($schema));
-    //
-    //     $this->expectException(SchemaException::class);
-    //     $this->expectExceptionMessage(
-    //         'A cyclic dependency was found when requesting a relations schema for role `employee`.'
-    //     );
-    //
-    //     (new Select($this->orm, Employee::class))->fetchData();
-    // }
-}
